+++ modules/tinymce/src/core/main/ts/content/SetContent.ts
@@ -5,9 +5,12 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
+import { Cell } from '@ephox/katamari';
+
 import Editor from '../api/Editor';
 import * as Rtc from '../Rtc';
 import { Content, SetContentArgs } from './ContentTypes';
+import { SetContentFormatter } from './EditorContent';
 
-export const setContent = (editor: Editor, content: Content, args: Partial<SetContentArgs> = {}): Content =>
-  Rtc.setContent(editor, content, args);
+export const setContent = (editor: Editor, content: Content, args: Partial<SetContentArgs> = {}, setCell: Cell<Record<string, SetContentFormatter>>): Content =>
+  Rtc.setContent(editor, content, args, setCell);